yap-macOS:svncom
git-svn-id: https://subversive.cims.nyu.edu/exact@753 a02fd672-c32b-43fa-a32f-4ef1cc386a6b

+++ trunk/progs/robotics/triangle/QuadTree.h
@@ -107,7 +107,7 @@ public:
 			return false;
 		}
 
-		for (int i = 0; i < cldrn.size(); ++i)
+		for (int i = 0; i < (int)cldrn.size(); ++i)
 		{
 			cldrn[i]->updateStatus();
 			insertNode(cldrn[i]);
@@ -133,7 +133,7 @@ public:
 			{
 				assert(b->status == Box::MIXED);
 
-				for (int i = 0; i < cldrn.size(); ++i)
+				for (int i = 0; i < (int)cldrn.size(); ++i)
 				{
 					cldrn[i]->updateStatus();
 					insertNode(cldrn[i]);
